requirements.txt
pillow==9.5.0
moviepy==1.0.3
requests>=2.25.0
numpy>=1.21.0
pyyaml>=6.0
ffmpeg-python>=0.2.0 
